requirements.txt
azure-cosmos==3.0.2
azure-functions==1.0.0a5
azure-functions-worker==1.0.0a6
certifi==2018.11.29
chardet==3.0.4
Django==2.1.3
grpcio==1.14.2
grpcio-tools==1.14.2
gunicorn==19.6.0
idna==2.7
numpy==1.15.4
protobuf==3.6.1
pytz==2018.7
requests==2.20.1
six==1.11.0
urllib3==1.24.1
python-dotenv